fix: refactor create spot state to custom hook
